run black

+++ GreyNsights/mpc.py
@@ -1,14 +1,15 @@
 import random
 
+# Use larger Prime number
 Q = 100103036233
 
 
-def encode(x, base=2, precision=16):
+def encode(x, base=2, precision=32):
     scale = base ** precision
     return x * scale
 
 
-def decode(x, base=2, precision=16):
+def decode(x, base=2, precision=32):
     correction = x < 0
     scale = base ** precision
     dividend = x // scale - correction
@@ -19,15 +20,12 @@ def decode(x, base=2, precision=16):
     return tensor
 
 
-# Add fixed precision with a larger Prime number
 def gen_shares(x, n):
 
     x = encode(x)
-
     shares = [random.randrange(Q) for worker in range(0, n - 1)]
 
     partial = 0
-
     for share in shares:
         partial += share
